+++ src/cdpx/primitives/profiler/__init__.py
@@ -1,7 +1,7 @@
 """Symfony Web Profiler collection and parsing primitives."""
 
 from .catalog import ALL_PANELS, PANEL_SOURCES
-from .collection import collect_profiler_report, fetch_panels, normalize_panels
+from .collection import collect_profiler_report, fetch_panels, fetch_profiler, normalize_panels
 from .parsers import parse_panel
 
 __all__ = [
@@ -9,6 +9,7 @@ __all__ = [
     "PANEL_SOURCES",
     "collect_profiler_report",
     "fetch_panels",
+    "fetch_profiler",
     "normalize_panels",
     "parse_panel",
 ]